[Misc]: move init_lmcache_engine() to v1 adapter (#1283)
* move init_lmcache_engine() to v1 adapter


* ready to branch


* try using a single config


* fix integration tests


---------
